requirements.txt

flask
flask-socketio
eventlet
torch
transformers
ultralytics
opencv-python
pygame
python-docx
Pillow
requests
